+++ packages/core-bridge/src/worker.rs
@@ -546,6 +546,7 @@ mod config {
         default_heartbeat_throttle_interval: Duration,
         max_activities_per_second: Option<f64>,
         max_task_queue_activities_per_second: Option<f64>,
+        max_concurrent_eager_activity_execution_size: usize,
         shutdown_grace_time: Option<Duration>,
         plugins: Vec<String>,
         storage_drivers: Vec<String>,
@@ -639,6 +640,9 @@ mod config {
                     self.max_task_queue_activities_per_second,
                 )
                 .maybe_max_worker_activities_per_second(self.max_activities_per_second)
+                .max_concurrent_eager_activity_execution_size(
+                    self.max_concurrent_eager_activity_execution_size,
+                )
                 .maybe_graceful_shutdown_period(self.shutdown_grace_time)
                 .plugins(
                     self.plugins